Welp
So a couple things. We have the host camera now working. The client camera is jiggling out of control and will not register the event of the input. Also I kinda killed motion. I am choosing not to fix it right now due to an automatic 'rail/track' type movement later.